chore: kmprevocationfactory reform
